remove old wifi manager
