+++ selfdrive/ui/layouts/settings/settings.py
@@ -7,8 +7,7 @@ from openpilot.selfdrive.ui.layouts.settings.device import DeviceLayout
 from openpilot.selfdrive.ui.layouts.settings.firehose import FirehoseLayout
 from openpilot.selfdrive.ui.layouts.settings.software import SoftwareLayout
 from openpilot.selfdrive.ui.layouts.settings.toggles import TogglesLayout
-from openpilot.system.ui.lib.wifi_manager import WifiManagerWrapper
-from openpilot.system.ui.lib.wifi_manager_v2 import WifiManager
+from openpilot.system.ui.lib.wifi_manager import WifiManager
 from openpilot.system.ui.lib.application import gui_app, FontWeight, MousePos
 from openpilot.system.ui.lib.text_measure import measure_text_cached
 from openpilot.system.ui.widgets.network import WifiManagerUI
