+++ system/ui/updater.py
@@ -7,7 +7,7 @@ from enum import IntEnum
 
 from openpilot.system.hardware import HARDWARE
 from openpilot.system.ui.lib.application import gui_app, FontWeight
-from openpilot.system.ui.lib.wifi_manager import WifiManagerWrapper
+from openpilot.system.ui.lib.wifi_manager import WifiManager
 from openpilot.system.ui.widgets import Widget
 from openpilot.system.ui.widgets.button import gui_button, ButtonStyle
 from openpilot.system.ui.widgets.label import gui_text_box, gui_label
@@ -43,7 +43,7 @@ class Updater(Widget):
     self.show_reboot_button = False
     self.process = None
     self.update_thread = None
-    self.wifi_manager = WifiManagerWrapper()
+    self.wifi_manager = WifiManager()
     self.wifi_manager_ui = WifiManagerUI(self.wifi_manager)
 
   def install_update(self):
